Inline rbjitglue into Ruby repository
When initially created, it seemed to make sense to keep the Ruby JIT's
'glue' code in a separate repository. The hope was to emphasize the
relativel ylow impact of the JIT compiler on the Ruby interpreter
overall.

However, further experience has found that this isn't a terribly
beneficial split for a number of reasons:

1. The JIT glue code is fairly tightly tied to the VM code, creating
   friction when working on new functionality. The distinction is
   particularly clear when comparing against the OMR submodule, which
   is dramatically more independent compared to the rbjitglue module.

2. The tight coupling leads to poor choices in engineering by keeping
   the module in a separate repo. Logically atomic changes become two
   commits, one in the VM and one in the glue.

I've talked about this in Issue 13. Similarly, I investigated a method
that would allow us to preserve the history of the submodule, but I've
come to the conclusion there's not enough value. The submodule history
still exists, but from this point forward, the JIT glue will be
developed on the trunk following branch inline.