Revive Inliner support.
Make the changes required to re-enable the inliner. The inliner now
peeks into the CALL_CACHE associated with a call to decide what to
inline, then guards the inlined bodies with serial number checks.

This means that for the inliner to successfully run, at the iseq must
have run at least once to have filled in the CALL_CACHE.