Use a single stack memory region for IlGen.
Turns out, I was being insufficiently careful in my useage of
TR::StackMemoryRegion. In order to tightly bound the lifetime of the map
created in genOptCaseDispatch, I tried to create a new stack memory
region, intending the memory of that map to be destroyed at the end of
that scope.

However, what I neglected to realize was that other usages of
trStackMemory() were occuring during the life of that region, and the
expectation was that they were going to live longer than just
genOptCaseDispatch. For example, the creation of entry stack for blocks
when calling genTarget.

In some circumstances, this was leading to bizzare failures where the
stack allocated data were being clobbered by other writes later.

See Issue #40 for one instance of this.

The resolution is to use a single stack memory region throughout the
ilgen process.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>