Memorize the correct method_entry on inlining.
An ongoing issue in the inliner has been the mixing of data that's
supposed to be constant (because inlining has occurred of a _particular_
method)  and data that mutates inside the CALL_CACHE (because it's a
cache!).

A test program like this:

    class Base
       attr_accessor :active

       def active?
          return true
       end

       def step
          @active = active? # (1)
       end

    end

    class ImplA < Base
       def active?
          super
       end
    end

    class ImplB < Base
       def active?
          super
       end
    end

    a = ImplA.new
    b = ImplB.new

    4.times {
       a.step
       b.step
    }

run with the inliner turns on sees the call to Base#active? get inlined to
whatever the last callee was. However, the next call fails the guard (because
of the a,b,a,b alternation -- nevertheless, the CALL_CACHE in Base#step is
updated.

Subsequently, when the inline body is dispatched next time, it uses the
method_entry from the call cache to setup the frame, which then fails the
call to super.

This patch fixes the issue by nailing the method_entry into the generated code.
This is fine, because we're already guarding the inlined code with the inline
guard.

Fixes Issue #59