Add TracePoint for basic operation redefinition.
This infrequent event can have an outsized performance impact by
eliminating a lot of the optimization that exists in the Ruby
interpreter.

In order to know when these sorts of things happen, this
patch adds a trace point that will fire when a basic operation is
redefined.

The event data can be accessed by calling #basic_operation_redefined,
which returns a hash of the class being modified and the basic operation
that's been redefined.

This patch may also be useful for JIT compilers that want to speculate
on basic operation definitions.

(cherry picked from commit 3aac74c3c2652ca291f2ed3555561bb27969af6b)