Migrate parts compilation logic to JIT DLL.
Rather than compile_f returning a body_info pointer and forcing the VM
to to manage the compilation logic, this change moves that logic inside
of compile_f, in preparation for asynchronous compilation.

This change also gets rid of some superflous functions originally
intended to provide sanity checks, however, their purpose is dead, so
the code cna be removed.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>